Rename org.thunlp.manage to org.thunlp.adjustment
Since most of the classes under this package is used to adjust the sentence of tagged words, use the term "adjustment" to represent them all.
